refactor(game): game 대신 dto를 생성하는 assembler 생성
